Move receipts tests to a single file.
